256/true colors: use shorter escape sequences
Most terminals limit the number of values in an SGR sequence, and the
limit can vary between terminals.

The 256 colors mode previously used one sequence with at least 6 values
(foreground and background of 256 colors), which is beyond the limit of
the Haiku terminal, and as a result it did not display the colors.

This is not a regression - we always produced such sequences.

Now we use one sequence for foreground, and one for background (and one
for additional global values, if given) - which does work in Haiku.

Similarly, the true color output was also split into two sequences.

There could be other places remaining which produce long escape
sequences, but for now all the default modes work out of the box also
on Haiku.